# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CareNear API",
        "description": "CareNear backend API — Plateforme de mise en relation patients / infirmiers à domicile.",
        "contact": {
            "email": "contact@carenear.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Local dev server"
        }
    ],
    "paths": {
...
        "/patient/profile": {
            "get": {
                "tags": [
                    "Patient Profile"
                ],
                "summary": "Get current patient profile",
                "description": "Returns the patient profile of the authenticated user including medical documents metadata. Returns null data if no profile exists.",
                "operationId": "4c99cd477f59f060a81a691a286b2348",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "birth_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1990-05-15",
                                                    "nullable": true
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "enum": [
                                                        "M",
                                                        "F",
                                                        "OTHER"
                                                    ],
                                                    "example": "M",
                                                    "nullable": true
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Nouakchott",
                                                    "nullable": true
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Ksar",
                                                    "nullable": true
                                                },
                                                "lat": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": true
                                                },
                                                "lng": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": true
                                                },
                                                "medical_notes": {
                                                    "type": "string",
                                                    "example": "Diabétique type 2",
                                                    "nullable": true
                                                },
                                                "medical_documents": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "index": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "filename": {
                                                                "type": "string",
                                                                "example": "1707123456_abc12345.pdf"
                                                            },
                                                            "url": {
                                                                "type": "string",
                                                                "example": "http://localhost/api/v1/patient/profile/documents/0/download"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "documents_count": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patient Profile"
                ],
                "summary": "Create or update patient profile (with optional medical documents upload)",
                "description": "PATIENT only. Uses multipart/form-data when uploading files. Maximum 5 files per request, 10 total. Accepted formats: PDF, JPG, PNG (max 5MB each).",
                "operationId": "2c86bb9619f407bff0ecfa78ef036f50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F",
                                            "OTHER"
                                        ],
                                        "example": "M"
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "Nouakchott"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 2000,
                                        "example": "Ksar, rue 42"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 18.086
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "double",
                                        "example": -15.978
                                    },
                                    "medical_notes": {
                                        "type": "string",
                                        "maxLength": 5000,
                                        "example": "Diabétique type 2, sous insuline"
                                    },
                                    "medical_files[]": {
                                        "description": "Medical documents (PDF/JPG/PNG, max 5MB each, max 5 files)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile created/updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — user is not a PATIENT"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patient/profile/documents/{index}": {
            "delete": {
                "tags": [
                    "Patient Profile"
                ],
                "summary": "Delete a specific medical document",
                "description": "PATIENT only. Deletes the medical document at the given index. Index is 0-based as returned by GET /patient/profile.",
                "operationId": "8cf70711102b6273459528266b72ed11",
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "description": "0-based index of the document to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Document deleted."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "No documents or invalid index"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Sanctum token as Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}